Refactor VaultBackupFlow component to use new password handling structure. Introduce VaultBackupWithoutPassword and VaultBackupWithPassword components for improved backup flow. Update routing and related imports accordingly. Add tasks for testing backup functionality.
