+++ clients/desktop/src/router.tsx
@@ -10,7 +10,7 @@ import { appPaths } from './navigation'
 import { OnboardingPage } from './onboarding/components/OnboardingPage'
 import { IncompleteOnboardingOnly } from './onboarding/IncompleteOnboardingOnly'
 import EditVaultPage from './pages/edItVault/EditVaultPage'
-import VaultBackupPage from './pages/edItVault/vaultBackupSettings/VaultBackupPage'
+import { VaultBackupPage } from './pages/edItVault/vaultBackupSettings/VaultBackupPage'
 import DeleteVaultPage from './pages/edItVault/vaultDeleteSettings/DeleteVaultPage'
 import VaultDetailsPage from './pages/edItVault/vaultDetailsSettings/VaultDetailsPage'
 import VaultRenamePage from './pages/edItVault/vaultRenameSettings/VaultRenamePage'
