+++ clients/desktop/src/vault/keygen/shared/VaultKeygenBackupFlow.tsx
@@ -3,7 +3,7 @@ import { BackupFastVault } from '../../setup/fast/backup/BackupFastVault'
 import { BackupSecureVault } from '../../setup/secure/backup/BackupSecureVault'
 import { useCurrentVault } from '../../state/currentVault'
 
-export const VaultBackupFlow = () => {
+export const VaultKeygenBackupFlow = () => {
   const vault = useCurrentVault()
 
   if (hasServerSigner(vault.signers)) {
